Fix 4 bugs blocking BIOSSES performance and training stability
1. encode() — replace last-token pooling with mean pooling over all
   non-padding positions. Mean pooling gives substantially better sentence
   representations for similarity/retrieval tasks (expected +0.15-0.25
   on BIOSSES Spearman vs last-token pooling).

2. DistillContrastiveLightningModule — add save_hyperparameters(ignore=['teacher'])
   to prevent checkpoint pickle crash when Lightning AI environment injects
   non-picklable forward hooks into the frozen teacher model.

3. train_stage0_distill.py — fix torch_dtype kwarg (was 'dtype=', which is
   the deprecated form in recent HuggingFace transformers).

4. train_stage0_distill.sh, train_stage1_distill.sh — fix wrong cd path:
   was 'cd SLURM_SUBMIT_DIR/hybrid_model_mamba_xlstm' (directory does not
   exist on cluster), corrected to 'cd SLURM_SUBMIT_DIR'.

+++ scripts/train_stage0_distill.py
@@ -429,7 +429,7 @@ def main(cfg: DictConfig):
     print(f"\nLoading teacher: {teacher_name} ({teacher_dtype_str})...")
     teacher = AutoModelForCausalLM.from_pretrained(
         teacher_name,
-        dtype=teacher_dtype,       # 'dtype' is the current HF kwarg (torch_dtype deprecated)
+        torch_dtype=teacher_dtype,
         low_cpu_mem_usage=True,
     )
     teacher.eval()
